Increased debugging and optimized a query path
